% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSquashingEM.R
\name{DataSquashingEM}
\alias{DataSquashingEM}
\title{HGZIPS - Data Squashing MGPS (optim)}
\usage{
DataSquashingEM(
  alpha1,
  beta1,
  alpha2,
  beta2,
  pi,
  squashedData,
  iteration,
  Loglik
)
}
\arguments{
\item{alpha1}{initial shape parameter value of the first gamma distribution for implementing the EM algprithm}

\item{beta1}{initial rate parameter value of the first gamma distribution for implementing the EM algprithm}

\item{alpha2}{initial shape parameter value of the second gamma distribution for implementing the EM algprithm}

\item{beta2}{initial rate parameter value of the second gamma distribution for implementing the EM algprithm}

\item{pi}{initial xxxxxxx?}

\item{squashedData}{the dataset of squashed N_ij, E_ij and weights. This dataset can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated parameters and their corresponding loglikelihood
}
\description{
This Data Squashing MGPS function.........
}
\seealso{

}
