% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSquashingHZINB.R
\name{DataSquashingHZINB}
\alias{DataSquashingHZINB}
\alias{parRangeCheck}
\title{Data Squashing HZINB function}
\usage{
parRangeCheck(N_ij, E_ij)
}
\arguments{
\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_klh}{initial probability value of all the alpha, beta, omega combinations for implementing the EM algprithm}

\item{N_ij_squashed}{a list of squashed datasets that include N_ij, E_ij and weights for each drug (j). This dataset list can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\description{
HGZIPS - Data Squashing HZINB (not assuming independence)
}
\details{
This Data Squashing HZINB function.........
}
