% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSquashingZINB.R
\name{DataSquashingZINB}
\alias{DataSquashingZINB}
\alias{LSE_R}
\title{HGZIPS - Data Squashing ZINB}
\usage{
LSE_R(vec)

DataSquashingZINB(alpha, beta, omega, squashedData, iteration, Loglik)
}
\arguments{
\item{alpha}{initial shape parameter value of the gamma distribution for implementing the EM algprithm}

\item{beta}{initial rate parameter value of the gamma distribution for implementing the EM algprithm}

\item{omega}{initial xxxxxxx?}

\item{squashedData}{the dataset of squashed N_ij, E_ij and weights. This dataset can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated parameters and their corresponding loglikelihood
DataSquashingZINB
}
\description{
This Data Squashing ZINB function.........
}
\seealso{

}
