% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\alias{observed_freqencies}
\alias{fitted_ZINB}
\alias{fitted_MGPS}
\alias{fitted_ZINB_two_gamma}
\alias{observed_freq_HZINB}
\alias{fitted_HZINB_one_gamma_independence}
\title{HGZIPS - Data Squashing HZINB (not assuming independence)}
\usage{
observed_freqencies(N)

fitted_ZINB(alpha, beta, omega, N, E)

fitted_MGPS(alpha1, alpha2, beta1, beta2, pi, N, E)

fitted_ZINB_two_gamma(alpha1, alpha2, beta1, beta2, pi, omega, N, E)

observed_freq_HZINB(N_ij)

fitted_HZINB_one_gamma_independence(
  posterior_a_j,
  posterior_b_j,
  posterior_omega_j,
  N_ij,
  E_ij
)

gof(fitted_prob, observed_freq)
}
\arguments{
\item{N}{1}

\item{alpha}{3}

\item{beta}{4}

\item{omega}{2}

\item{E}{2}

\item{alpha1}{5}

\item{alpha2}{6}

\item{beta1}{1}

\item{beta2}{2}

\item{pi}{3}

\item{N_ij}{3}

\item{posterior_a_j}{2}

\item{posterior_b_j}{2}

\item{posterior_omega_j}{23}

\item{E_ij}{4}

\item{fitted_prob}{3}

\item{observed_freq}{4}
}
\value{
observed frequencies, estimated frequencies and goodness of fit

observed frequencies
observed_frequencies

ZINB fitted values
fitted_ZINB

fitted MGPS
fitted MGPS

fitted ZINB two gamma
fitted ZINB two gamma

observed_freq_HZINB
observed_freq_HZINB

fitted_HZINB_one_gamma_independence
fitted_HZINB_one_gamma_independence

gof
gof
}
\description{
This Data Squashing HZINB function.........
}
\seealso{

}
