% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HZINB_one_gamma.R
\name{HZINB}
\alias{HZINB}
\alias{parRangeCheck}
\alias{HZINB_one_gamma}
\title{HGZIPS - HZINB (not assuming independence)}
\usage{
parRangeCheck(N_ij, E_ij)

HZINB_one_gamma(
  grid_a,
  grid_b,
  grid_omega,
  init_pi_klh,
  N_ij,
  E_ij,
  iteration,
  Loglik
)
}
\arguments{
\item{N_ij}{matrix of N_ij, i = AE, j = drugs}

\item{E_ij}{matrix of E_ij, i = AE, j = drugs}

\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_klh}{initial probability value of all the alpha, beta, omega combinations for implementing the EM algprithm}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
}
\description{
This HZINB function.........
}
\seealso{

}
