% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HZINB_ind_two_gamma.R
\name{HZINB_ind_two_gamma}
\alias{HZINB_ind_two_gamma}
\alias{grid_HZINB_two_gamma}
\title{HGZIPS - HZINB with two gamma components (assuming independence)}
\usage{
grid_HZINB_two_gamma(a_j, b_j, omega_j, K1, L1, K2, L2, M, H)

HZINB_ind_two_gamma(
  grid_a1,
  grid_a2,
  grid_b1,
  grid_b2,
  grid_pi,
  grid_omega,
  init_pi_k1,
  init_pi_l1,
  init_pi_k2,
  init_pi_l2,
  init_pi_m,
  init_pi_h,
  dataset,
  iteration,
  Loglik = FALSE,
  zeroes = FALSE,
  N_star = 1
)
}
\arguments{
\item{grid_a1}{alpha1 value grid}

\item{grid_a2}{alpha2 value grid}

\item{grid_b1}{beta1 value grid}

\item{grid_b2}{beta2 value grid}

\item{grid_pi}{pi value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_k1}{initial probability of each alpha1 value for implementing the EM algorithm}

\item{init_pi_l1}{initial probability of each beta1 value for implementing the EM algorithm}

\item{init_pi_k2}{initial probability of each alpha2 value for implementing the EM algorithm}

\item{init_pi_l2}{initial probability of each beta2 value for implementing the EM algorithm}

\item{init_pi_m}{initial probability of each pi value for implementing the EM algorithm}

\item{init_pi_h}{initial probability of each omega value for implementing the EM algprithm}

\item{dataset}{a list of squashed datasets that include N_ij, E_ij and weights for each drug (j). This dataset list can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}

\item{zeroes}{A logical scalar specifying if zero counts should be included.}

\item{N_star}{the minimum Nij count size to be used for hyperparameter estimation. If zeroes are included in Nij vector, please set N_star = NULL}
}
\value{
\code{grid_HZINB} build a suitable grid of a_j, b_j, and omega_j for implementing HZINB
grid_HZINB

\code{HZINB_ind_two_gamma} a list of estimated probability of each alpha1, beta1, alpha2, beta2, pi, omega combination and their corresponding loglikelihood (optional)
\itemize{
\item \code{theta_EM} Estimate of hyperparameters for each EM iteration
\item \code{llh} logliklihood for each EM iteration (optional)
}
}
\description{
This \code{HZINB_ind_two_gamma} function finds hyperparameter estimates by implementing the Expectation-Maximization (EM) algorithm and hierarchical zero-inflated negative binomial model with two gamma components.
}
