% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HZINB.R
\name{HZINB_one_gamma}
\alias{HZINB_one_gamma}
\alias{parRangeCheck}
\alias{grid_HZINB}
\title{HGZIPS - HZINB (not assuming independence)}
\usage{
parRangeCheck(N_ij, E_ij)

grid_HZINB(a_j, b_j, omega_j, K, L, H)

HZINB_one_gamma(
  grid_a,
  grid_b,
  grid_omega,
  init_pi_klh,
  dataset,
  iteration,
  Loglik = FALSE,
  zeroes = FALSE,
  N_star = 1
)
}
\arguments{
\item{N_ij}{matrix of N_ij values, i - AE, j - drugs}

\item{E_ij}{matrix of E_ij values, corresponding to N_ij}

\item{a_j}{alpha for gamma distribution}

\item{b_j}{beta for gamma distribution}

\item{omega_j}{proportion}

\item{K}{number of a_j in grid}

\item{L}{number of b_j in grid}

\item{H}{number of omega_j in grid}

\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_klh}{initial probability value of all the alpha, beta, omega combinations for implementing the EM algprithm}

\item{dataset}{a list of squashed datasets that include N_ij, E_ij and weights for each drug (j). This dataset list can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}

\item{zeroes}{A logical scalar specifying if zero counts should be included.}

\item{N_star}{the minimum Nij count size to be used for hyperparameter estimation. If zeroes are included in Nij vector, please set N_star = NULL}
}
\value{
\code{parRangeCheck} the estimated a_j, b_j and omega_j for each drug (j)
parRangeCheck

\code{grid_HZINB} build a suitable grid of a_j, b_j, and omega_j for implementing HZINB
grid_HZINB

\code{HZINB_one_gamma} a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
\itemize{
\item \code{theta_EM} Estimate of hyperparameters for each EM iteration
\item \code{llh} logliklihood for each EM iteration (optional)
}
HZINB_one_gamma
}
\description{
This \code{HZINB_one_gamma} function finds hyperparameter estimates by implementing the Expectation-Maximization (EM) algorithm and hierarchical zero-inflated negative binomial model with one gamma component.
}
