% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{log Sum Exp}
\usage{
log_sum_exp(u, v)
}
\arguments{
\item{u}{first number}

\item{v}{second number}
}
\value{
logged sum of exponential of u and v
}
\description{
log Sum Exp
}
\seealso{

}
