% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{LogSumExp}
\alias{LogSumExp}
\alias{log_sum_exp}
\alias{LSE_R}
\title{log Sum Exp}
\usage{
log_sum_exp(u, v)

LSE_R(vec)
}
\arguments{
\item{u}{first number}

\item{v}{second number}

\item{vec}{vector to be summed}
}
\value{
logged sum of exponential of u and v

the logged sum of the exponentials of all values in vector
}
\description{
log Sum Exp

log Sum Exp of vector
}
