% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSE_R.R
\name{LSE_R}
\alias{LSE_R}
\title{log Sum Exp of vector}
\usage{
LSE_R(vec)
}
\arguments{
\item{vec}{vector to be summed}
}
\value{
the logged sum of the exponentials of all values in vector
}
\description{
log Sum Exp of vector
}
\seealso{

}
