% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSquashingEM_uni_opt.R,
%   R/MGPS_uniroot_optimization.R
\name{MGPS(uni-opt)}
\alias{MGPS(uni-opt)}
\alias{ProfileLogLik}
\alias{DSMGPS_uniroot_optimization}
\alias{MGPS_uniroot_optimization}
\title{HGZIPS - Data Squashing MGPS (uniroot-optimization)}
\usage{
ProfileLogLik(alpha, Tij, N, E)

ProfileLogLik(alpha, Tij, N, E)
}
\arguments{
\item{N}{vector of N_ij}

\item{E}{vector of E_ij}

\item{weight}{weights (vector) of each N_ij value after data squashing. This data can be generated by the rawProcessing function in this package.}

\item{alpha.par}{initial shape parameter vector of the two gamma distributions for implementing the EM algprithm}

\item{beta.par}{initial rate parameter vector of the two gamma distributions for implementing the EM algprithm}

\item{pi.par}{initial xxxxxxx?}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\description{
This Data Squashing MGPS function.........

HGZIPS - MGPS (uniroot-optimization)
}
\details{
This MGPS function.........
}
\seealso{



}
