% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPS_uni_opt.R
\name{MGPS_uni_opt}
\alias{MGPS_uni_opt}
\alias{ProfileLogLik}
\title{HGZIPS - MGPS (uniroot-optimization)}
\usage{
ProfileLogLik(alpha, Tij, weight, N, E)

MGPS_uni_opt(alpha.par, beta.par, pi.par, N, E, weight, iterations, Loglik)
}
\arguments{
\item{weight}{set weight = rep(1, length(N)) if N and E are not squashed data, or input the weight vector corresponding to the squashed Nij vector.}

\item{N}{squashed N_ij data (vector). This data can be generated by the rawProcessing function in this package.}

\item{E}{squashed E_ij data (vector). This data can be generated by the rawProcessing function in this package.}

\item{alpha.par}{initial shape parameter vector of the two gamma distributions for implementing the EM algprithm}

\item{beta.par}{initial rate parameter vector of the two gamma distributions for implementing the EM algprithm}

\item{pi.par}{initial xxxxxxx?}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}

\item{iteration}{number of EM algorithm iterations to run}
}
\value{
a list of estimated parameters and their corresponding loglikelihood
}
\description{
This MGPS function.........
}
\seealso{

}
