% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSquashingHZINB.R, R/HZINB_one_gamma.R
\name{parRangeCheck}
\alias{parRangeCheck}
\title{Create a complete ggplot appropriate to a particular data type}
\usage{
parRangeCheck(N_ij, E_ij)

parRangeCheck(N_ij, E_ij)
}
\arguments{
\item{object}{an object, whose class will determin the behaviour of autoplot}

\item{...other}{auguments passed to specific methods}
}
\value{
a ggplot object

a ggplot object
}
\description{
\code{autoplot} uses ggplot2 to draw a particular plot for an object of a particular class in a single command.
This defines the S3 generic that other classes and packages can extend.

\code{autoplot} uses ggplot2 to draw a particular plot for an object of a particular class in a single command.
This defines the S3 generic that other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}

\code{\link{ggplot}} and \code{\link{fortify}}
}
