% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_ZINB.R
\name{post_mean_lambda_ZINB}
\alias{post_mean_lambda_ZINB}
\title{Create a complete ggplot appropriate to a particular data type}
\usage{
post_mean_lambda_ZINB(alpha, beta, N, E)
}
\arguments{
\item{object}{an object, whose class will determin the behaviour of autoplot}

\item{...other}{auguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
\code{autoplot} uses ggplot2 to draw a particular plot for an object of a particular class in a single command.
This defines the S3 generic that other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}
}
