% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior}
\alias{posterior}
\alias{posterior_abomega}
\alias{posterior_HZINB}
\alias{post_mean_lambda_ZINB}
\alias{post_mean_loglambda_ZINB}
\alias{Posteror_MGPS}
\alias{Posteror_MGPS_log}
\title{HGZIPS - calculating the posterior value of HZINB a_j, b_j and omega_j}
\usage{
posterior_abomega(
  grid_a,
  grid_b,
  grid_omega,
  pi_klh_final_a_j,
  pi_klh_final_b_j,
  pi_klh_final_omega_j,
  N_ij,
  E_ij
)

posterior_HZINB(posterior_a_j, posterior_b_j, posterior_omega_j, N_ij, E_ij)

post_mean_lambda_ZINB(alpha, beta, N, E)

post_mean_loglambda_ZINB(alpha, beta, N, E)

Posteror_MGPS(alpha1, beta1, alpha2, beta2, pi, N, E)

Posteror_MGPS_log(alpha1, beta1, alpha2, beta2, pi, N, E)
}
\arguments{
\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{pi_klh_final_a_j}{final estimation of the probability of each alpha value, produced by the EM algorithm}

\item{pi_klh_final_b_j}{final estimation of the probability of each beta value, produced by the EM algorithm}

\item{pi_klh_final_omega_j}{final estimation of the probability of each omega value, produced by the EM algorithm}

\item{N_ij}{matrix of N_ij, i = AE, j = drugs}

\item{E_ij}{matrix of E_ij, i = AE, j = drugs}
}
\value{
a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)

a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)

posterior mean of logged lambda

a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)

a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
}
\description{
HGZIPS - calculating the posterior value of HZINB a_j, b_j and omega_j
}
\seealso{

}
