% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_abomega_HZINB.R
\name{posterior_abomega}
\alias{posterior_abomega}
\title{Create a complete ggplot appropriate to a particular data type}
\usage{
posterior_abomega(
  grid,
  pi_klh_final_a_j,
  pi_klh_final_b_j,
  pi_klh_final_omega_j,
  N_ij,
  E_ij
)
}
\arguments{
\item{object}{an object, whose class will determin the behaviour of autoplot}

\item{...other}{auguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
\code{autoplot} uses ggplot2 to draw a particular plot for an object of a particular class in a single command.
This defines the S3 generic that other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}
}
