% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_HZINB.R
\name{posteriorHZINB}
\alias{posteriorHZINB}
\title{HGZIPS - HZINB posterior lambda}
\usage{
posteriorHZINB(posterior_a_j, posterior_b_j, posterior_omega_j, N_ij, E_ij)
}
\arguments{
\item{posterior_a_j}{the posterior average of HZINB a_j, can be produced by the posterior_abomega_HZINB function in the package}

\item{posterior_b_j}{the posterior average of HZINB b_j, can be produced by the posterior_abomega_HZINB function in the package}

\item{posterior_omega_j}{the posterior average of HZINB b_j, can be produced by the posterior_abomega_HZINB function in the package}

\item{N_ij}{matrix of N_ij, i = AE, j = drugs}

\item{E_ij}{matrix of E_ij, i = AE, j = drugs}
}
\value{
the posterior mean of the lambda in HZINB
}
\description{
This HZINB function.........
}
\seealso{

}
