% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteror_MGPS.R
\name{Posteror_MGPS}
\alias{Posteror_MGPS}
\title{HGZIPS - MGPS posterior lambda}
\usage{
Posteror_MGPS(alpha1, beta1, alpha2, beta2, pi, N, E)
}
\arguments{
\item{alpha1}{the final estimation of alpha1 in ZINB}

\item{beta1}{the final estimation of beta1 in ZINB}

\item{alpha2}{the final estimation of alpha2 in ZINB}

\item{beta2}{the final estimation of beta2 in ZINB}

\item{pi}{the final estimation of pi in ZINB}

\item{N}{vector of Nij values}

\item{E}{vector of Eij values}
}
\value{
the posterior mean of the lambda in HZINB
}
\description{
This HZINB function.........
}
\seealso{

}
