% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawProcessing.R
\name{rawProcessing}
\alias{rawProcessing}
\title{HGZIPS - rawProcessing}
\usage{
rawProcessing(rawdata, frequency_threshold, data.squashing = FALSE)
}
\arguments{
\item{rawdata}{rawdata input}

\item{frequency_threshold}{minimum......}

\item{data.squashing}{whether to conduct data squashing or not (TRUE or FALSE)}
}
\value{
a list of objects after raw data processing
}
\description{
This function.........
}
\seealso{

}
