% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB_one_gamma.R
\name{ZINB_one_gamma}
\alias{ZINB_one_gamma}
\title{HGZIPS - ZINB (one gamma)}
\usage{
ZINB_one_gamma(alpha, beta, omega, N, E, weight, iteration, Loglik)
}
\arguments{
\item{alpha}{initial shape parameter value of the gamma distribution for implementing the EM algprithm}

\item{beta}{initial rate parameter value of the gamma distribution for implementing the EM algprithm}

\item{omega}{initial xxxxxxx?}

\item{N}{vector of Nij values}

\item{E}{vector of Eij values}

\item{weight}{set weight = rep(1, length(N)) if N and E are not squashed data, or input the weight vector corresponding to the squashed Nij vector.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated parameters and their corresponding loglikelihood
DataSquashingZINB
}
\description{
This Data Squashing ZINB function.........
}
\seealso{

}
