% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB_two_gamma.R
\name{ZINB_two_gamma}
\alias{ZINB_two_gamma}
\title{HGZIPS - ZINB (mixture of two gamma)}
\usage{
ZINB_two_gamma(
  alpha1,
  beta1,
  alpha2,
  beta2,
  pi,
  omega,
  N,
  E,
  weight,
  iteration,
  Loglik
)
}
\arguments{
\item{alpha1}{initial shape parameter value of the first gamma distribution for implementing the EM algprithm}

\item{beta1}{initial rate parameter value of the first gamma distribution for implementing the EM algprithm}

\item{alpha2}{initial shape parameter value of the second gamma distribution for implementing the EM algprithm}

\item{beta2}{initial rate parameter value of the second gamma distribution for implementing the EM algprithm}

\item{pi}{initial xxxxxxx?}

\item{omega}{initial xxxxxxx?}

\item{N}{vector of Nij values}

\item{E}{vector of Eij values}

\item{weight}{set weight = rep(1, length(N)) if N and E are not squashed data, or input the weight vector corresponding to the squashed Nij vector.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated parameters and their corresponding loglikelihood
}
\description{
This ZINB function.........
}
\seealso{

}
