% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HZINB_independence.R
\name{HZINB_independence}
\alias{HZINB_independence}
\title{HGZIPS - HZINB ( assuming independence)}
\usage{
HZINB_independence(
  grid_a,
  grid_b,
  grid_omega,
  init_pi_k,
  init_pi_l,
  init_pi_h,
  dataset,
  iteration,
  Loglik
)
}
\arguments{
\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_k}{initial probability of each alpha value for implementing the EM algorithm}

\item{init_pi_l}{inital probability of each beta value for implementing the EM algorithm}

\item{init_pi_h}{initial probability of each omega value for implementing the EM algprithm}

\item{dataset}{a list of squashed datasets that include N_ij, E_ij and weights for each drug (j). This dataset list can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
HZINB_independence
}
\description{
This HZINB function.........
}
\seealso{

}
