% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HZINB.R
\name{HZINB_one_gamma}
\alias{HZINB_one_gamma}
\alias{parRangeCheck}
\alias{grid_HZINB}
\title{HGZIPS - HZINB (not assuming independence)}
\usage{
parRangeCheck(N_ij, E_ij)

grid_HZINB(a_j, b_j, omega_j, K, L, H)

HZINB_one_gamma(
  grid_a,
  grid_b,
  grid_omega,
  init_pi_klh,
  dataset,
  iteration,
  Loglik
)
}
\arguments{
\item{N_ij}{matrix of N_ij values, i - AE, j - drugs}

\item{E_ij}{matrix of E_ij values, corresponding to N_ij}

\item{a_j}{alpha for gamma distribution}

\item{b_j}{beta for gamma distribution}

\item{omega_j}{proportion}

\item{K}{number of a_j in grid}

\item{L}{number of b_j in grid}

\item{H}{number of omega_j in grid}

\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{init_pi_klh}{initial probability value of all the alpha, beta, omega combinations for implementing the EM algprithm}

\item{dataset}{a list of squashed datasets that include N_ij, E_ij and weights for each drug (j). This dataset list can be generated by the rawProcessing function in this package.}

\item{iteration}{number of EM algorithm iterations to run}

\item{Loglik}{whether to return the loglikelihood of each iteration or not (TRUE or FALSE)}
}
\value{
check the range of a_j, b_j, and omega_j for the dataset

a suitable grid of a_j, b_j, and omega_j for implementing HZINB

a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
}
\description{
This HZINB function.........
}
