% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_ZINB.R
\name{post_mean_lambda_ZINB}
\alias{post_mean_lambda_ZINB}
\title{HGZIPS - ZINB posterior lambda}
\usage{
post_mean_lambda_ZINB(alpha, beta, N, E)
}
\arguments{
\item{alpha}{the final estimation of alpha in ZINB}

\item{beta}{the final estimation of beta in ZINB}

\item{N}{vector of Nij values}

\item{E}{vector of Eij values}
}
\value{
the posterior mean of the lambda in ZINB
}
\description{
This HZINB function.........
}
\seealso{

}
