% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_abomega_HZINB.R
\name{posterior_abomega}
\alias{posterior_abomega}
\title{HGZIPS - calculating the posterior value of HZINB a_j, b_j and omega_j}
\usage{
posterior_abomega(
  grid_a,
  grid_b,
  grid_omega,
  pi_klh_final_a_j,
  pi_klh_final_b_j,
  pi_klh_final_omega_j,
  N_ij,
  E_ij
)
}
\arguments{
\item{grid_a}{alpha value grid}

\item{grid_b}{beta value grid}

\item{grid_omega}{omega value grid}

\item{pi_klh_final_a_j}{final estimation of the probability of each alpha value, produced by the EM algorithm}

\item{pi_klh_final_b_j}{final estimation of the probability of each beta value, produced by the EM algorithm}

\item{pi_klh_final_omega_j}{final estimation of the probability of each omega value, produced by the EM algorithm}

\item{N_ij}{matrix of N_ij, i = AE, j = drugs}

\item{E_ij}{matrix of E_ij, i = AE, j = drugs}
}
\value{
a list of estimated probability of each alpha, beta, omega combination and their corresponding loglikelihood (optional)
}
\description{
HGZIPS - calculating the posterior value of HZINB a_j, b_j and omega_j
}
\seealso{

}
