% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteror_MGPS.R
\name{Posteror_MGPS}
\alias{Posteror_MGPS}
\title{Create a complete ggplot appropriate to a particular data type}
\usage{
Posteror_MGPS(alpha1, beta1, alpha2, beta2, pi, N, E)
}
\arguments{
\item{object}{an object, whose class will determin the behaviour of autoplot}

\item{...other}{auguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
\code{autoplot} uses ggplot2 to draw a particular plot for an object of a particular class in a single command.
This defines the S3 generic that other classes and packages can extend.
}
\seealso{
\code{\link{ggplot}} and \code{\link{fortify}}
}
